proper arithmetic operator
`==` is a string operator. It happens to work fine in this case, but I'd recommend to use the proper arithmetic operator `-eq` instead.

Also, why not?
```sh
contains_element "${guard}" "${unique_guards[@]}" && \
    echo "Header guard '${guard}' is not unique!"
```
I did not test it, but the logic seems to be the same. If the function returns zero, echo message.